Add missing properties for error response object
The error response from Payload does not reflect the error response in the OpenAPI due to missing properties. This means that if anyone wants to use i.e openapi-fetch, they won't be able to use those properties or they will have to override the types.

I have added the missing properties and set them to optional to make sure it's backwards compatible. I tested the old schema vs the new schema with different error objects.